- fnc and cmd can now be grouped in "modules" - unified require paths - got rid of half the aysnc functions - better logging config + mail - added division2 clan xp rankings - temporary removed raids, will come back as events

+++ cmd/src/prune.js
@@ -1,5 +1,5 @@
-const CON = require('../../src/const.json');
-const fnc = require('../../fnc');
+const CON = require(`${require.main.path}/src/const.json`);
+const fnc = require(`${require.main.path}/fnc`);
 
 module.exports = {
 	description: 'Deletes a number of messages.',
@@ -14,7 +14,7 @@ module.exports = {
 		args[0] = parseInt(args[0]) + 1;
 		if (isNaN(args[0]) || args[0] < 2 || args[0] > 100) {
 			//if (message.channel.type === 'text') message.delete();
-			return fnc.replyExt(message, 'you need to input a number between 1 and 99', { color: CON.TEXTCLR.WARN });
+			return fnc.replyWarn(message, 'you need to input a number between 1 and 99');
 		}
 
 		if (message.channel.type === 'text') {
@@ -22,9 +22,14 @@ module.exports = {
 				await message.channel.bulkDelete(args[0], true);
 			}
 			catch (e) {
-				message.client.logger.warn(`Couldn't prune ${message.channel.name}:\n${e.stack}`);
-				message.fnc.replyExt(message, 'there was an error trying to prune messages in this channel', { color: CON.TEXTCLR.ERROR });
-				message.delete();
+				if (e.name === 'DiscordAPIError' && e.message === 'Missing Permissions' || e.message === 'Missing Access') {
+					message.channel.send(`${message.guild.owner}, i don't have permission to delete messages here!`);
+				}
+				else {
+					message.client.logger.warn(`Couldn't prune ${message.channel.name}:\n${e.stack}`);
+					fnc.replyWarn(message, 'there was an error trying to prune messages in this channel');
+					message.delete();
+				}
 			}
 		}
 		else if (message.channel.type === 'dm') {
